Create Traveling to Acadia
Imagine you work at a travel agency and want to inform your customers of the best time to visit one of your favorite vacation locations, Acadia, Maine. In this project,   used flower bloom data, and flight information to recommend the best time of year for someone to make a trip to Maine.